translate-pipeline/requirements.txt
google-cloud-translate==3.2.1